+++ crates/nu_plugin_polars/src/dataframe/core/save/arrow.rs
@@ -2,8 +2,8 @@ use std::{fs::File, path::Path};
 
 use nu_plugin::EvaluatedCall;
 use nu_protocol::{ShellError, Span};
-use polars::prelude::ParquetWriter;
-use polars_io::parquet::write::ParquetWriteOptions;
+use polars::prelude::{IpcWriter, SerWriter};
+use polars_io::ipc::IpcWriterOptions;
 
 use crate::values::{NuDataFrame, NuLazyFrame};
 
@@ -16,7 +16,7 @@ pub(crate) fn command_lazy(
     file_span: Span,
 ) -> Result<(), ShellError> {
     lazy.to_polars()
-        .sink_parquet(file_path, ParquetWriteOptions::default())
+        .sink_ipc(file_path, IpcWriterOptions::default())
         .map_err(|e| polars_file_save_error(e, file_span))
 }
 
@@ -25,16 +25,16 @@ pub(crate) fn command_eager(
     file_path: &Path,
     file_span: Span,
 ) -> Result<(), ShellError> {
-    let file = File::create(file_path).map_err(|e| ShellError::GenericError {
-        error: "Error with file name".into(),
-        msg: e.to_string(),
+    let mut file = File::create(file_path).map_err(|e| ShellError::GenericError {
+        error: format!("Error with file name: {e}"),
+        msg: "".into(),
         span: Some(file_span),
         help: None,
         inner: vec![],
     })?;
-    let mut polars_df = df.to_polars();
-    ParquetWriter::new(file)
-        .finish(&mut polars_df)
+
+    IpcWriter::new(&mut file)
+        .finish(&mut df.to_polars())
         .map_err(|e| ShellError::GenericError {
             error: "Error saving file".into(),
             msg: e.to_string(),
@@ -46,17 +46,17 @@ pub(crate) fn command_eager(
 }
 
 #[cfg(test)]
-pub(crate) mod test {
+pub mod test {
 
-    use crate::eager::save::test::{test_eager_save, test_lazy_save};
+    use crate::core::save::test::{test_eager_save, test_lazy_save};
 
     #[test]
-    pub fn test_parquet_eager_save() -> Result<(), Box<dyn std::error::Error>> {
-        test_eager_save("parquet")
+    pub fn test_arrow_eager_save() -> Result<(), Box<dyn std::error::Error>> {
+        test_eager_save("arrow")
     }
 
     #[test]
-    pub fn test_parquet_lazy_save() -> Result<(), Box<dyn std::error::Error>> {
-        test_lazy_save("parquet")
+    pub fn test_arrow_lazy_save() -> Result<(), Box<dyn std::error::Error>> {
+        test_lazy_save("arrow")
     }
 }
