+++ crates/nu_plugin_polars/src/dataframe/core/save/ndjson.rs
@@ -1,9 +1,9 @@
-use std::{fs::File, path::Path};
+use std::{fs::File, io::BufWriter, path::Path};
 
 use nu_plugin::EvaluatedCall;
 use nu_protocol::{ShellError, Span};
-use polars::prelude::{IpcWriter, SerWriter};
-use polars_io::ipc::IpcWriterOptions;
+use polars::prelude::{JsonWriter, SerWriter};
+use polars_io::json::JsonWriterOptions;
 
 use crate::values::{NuDataFrame, NuLazyFrame};
 
@@ -16,7 +16,7 @@ pub(crate) fn command_lazy(
     file_span: Span,
 ) -> Result<(), ShellError> {
     lazy.to_polars()
-        .sink_ipc(file_path, IpcWriterOptions::default())
+        .sink_json(file_path, JsonWriterOptions::default())
         .map_err(|e| polars_file_save_error(e, file_span))
 }
 
@@ -25,15 +25,16 @@ pub(crate) fn command_eager(
     file_path: &Path,
     file_span: Span,
 ) -> Result<(), ShellError> {
-    let mut file = File::create(file_path).map_err(|e| ShellError::GenericError {
+    let file = File::create(file_path).map_err(|e| ShellError::GenericError {
         error: format!("Error with file name: {e}"),
         msg: "".into(),
         span: Some(file_span),
         help: None,
         inner: vec![],
     })?;
+    let buf_writer = BufWriter::new(file);
 
-    IpcWriter::new(&mut file)
+    JsonWriter::new(buf_writer)
         .finish(&mut df.to_polars())
         .map_err(|e| ShellError::GenericError {
             error: "Error saving file".into(),
@@ -42,21 +43,21 @@ pub(crate) fn command_eager(
             help: None,
             inner: vec![],
         })?;
+
     Ok(())
 }
 
 #[cfg(test)]
 pub mod test {
-
-    use crate::eager::save::test::{test_eager_save, test_lazy_save};
+    use crate::core::save::test::{test_eager_save, test_lazy_save};
 
     #[test]
     pub fn test_arrow_eager_save() -> Result<(), Box<dyn std::error::Error>> {
-        test_eager_save("arrow")
+        test_eager_save("ndjson")
     }
 
     #[test]
     pub fn test_arrow_lazy_save() -> Result<(), Box<dyn std::error::Error>> {
-        test_lazy_save("arrow")
+        test_lazy_save("ndjson")
     }
 }
