+++ crates/nu_plugin_polars/src/dataframe/core/save/parquet.rs
@@ -1,9 +1,9 @@
-use std::{fs::File, io::BufWriter, path::Path};
+use std::{fs::File, path::Path};
 
 use nu_plugin::EvaluatedCall;
 use nu_protocol::{ShellError, Span};
-use polars::prelude::{JsonWriter, SerWriter};
-use polars_io::json::JsonWriterOptions;
+use polars::prelude::ParquetWriter;
+use polars_io::parquet::write::ParquetWriteOptions;
 
 use crate::values::{NuDataFrame, NuLazyFrame};
 
@@ -16,7 +16,7 @@ pub(crate) fn command_lazy(
     file_span: Span,
 ) -> Result<(), ShellError> {
     lazy.to_polars()
-        .sink_json(file_path, JsonWriterOptions::default())
+        .sink_parquet(file_path, ParquetWriteOptions::default())
         .map_err(|e| polars_file_save_error(e, file_span))
 }
 
@@ -26,16 +26,15 @@ pub(crate) fn command_eager(
     file_span: Span,
 ) -> Result<(), ShellError> {
     let file = File::create(file_path).map_err(|e| ShellError::GenericError {
-        error: format!("Error with file name: {e}"),
-        msg: "".into(),
+        error: "Error with file name".into(),
+        msg: e.to_string(),
         span: Some(file_span),
         help: None,
         inner: vec![],
     })?;
-    let buf_writer = BufWriter::new(file);
-
-    JsonWriter::new(buf_writer)
-        .finish(&mut df.to_polars())
+    let mut polars_df = df.to_polars();
+    ParquetWriter::new(file)
+        .finish(&mut polars_df)
         .map_err(|e| ShellError::GenericError {
             error: "Error saving file".into(),
             msg: e.to_string(),
@@ -43,21 +42,21 @@ pub(crate) fn command_eager(
             help: None,
             inner: vec![],
         })?;
-
     Ok(())
 }
 
 #[cfg(test)]
-pub mod test {
-    use crate::eager::save::test::{test_eager_save, test_lazy_save};
+pub(crate) mod test {
+
+    use crate::core::save::test::{test_eager_save, test_lazy_save};
 
     #[test]
-    pub fn test_arrow_eager_save() -> Result<(), Box<dyn std::error::Error>> {
-        test_eager_save("ndjson")
+    pub fn test_parquet_eager_save() -> Result<(), Box<dyn std::error::Error>> {
+        test_eager_save("parquet")
     }
 
     #[test]
-    pub fn test_arrow_lazy_save() -> Result<(), Box<dyn std::error::Error>> {
-        test_lazy_save("ndjson")
+    pub fn test_parquet_lazy_save() -> Result<(), Box<dyn std::error::Error>> {
+        test_lazy_save("parquet")
     }
 }
