+++ crates/nu_plugin_polars/src/dataframe/core/shape.rs
