+++ crates/nu_plugin_polars/src/dataframe/core/summary.rs
