+++ crates/nu_plugin_polars/src/dataframe/macro_commands.rs
@@ -1,14 +1,15 @@
-/// Definition of multiple lazyframe commands using a macro rule
-/// All of these commands have an identical body and only require
-/// to have a change in the name, description and function
-use crate::dataframe::values::{Column, NuDataFrame, NuLazyFrame};
-use crate::values::CustomValueSupport;
-use crate::PolarsPlugin;
-use nu_plugin::{EngineInterface, EvaluatedCall, PluginCommand};
-use nu_protocol::{Category, Example, LabeledError, PipelineData, Signature, Span, Type, Value};
-
+#[macro_export]
 macro_rules! lazy_command {
     ($command: ident, $name: expr, $desc: expr, $examples: expr, $func: ident, $test: ident) => {
+        use nu_plugin::{EngineInterface, EvaluatedCall, PluginCommand};
+        use nu_protocol::{Category, Example, LabeledError, PipelineData, Signature, Type};
+        /// Definition of multiple lazyframe commands using a macro rule
+        /// All of these commands have an identical body and only require
+        /// to have a change in the name, description and function
+        use $crate::dataframe::values::NuLazyFrame;
+        use $crate::values::CustomValueSupport;
+        use $crate::PolarsPlugin;
+
         #[derive(Clone)]
         pub struct $command;
 
@@ -55,8 +56,8 @@ macro_rules! lazy_command {
         #[cfg(test)]
         mod $test {
             use super::*;
-            use crate::test::test_polars_plugin_command;
             use nu_protocol::ShellError;
+            use $crate::test::test_polars_plugin_command;
 
             #[test]
             fn test_examples() -> Result<(), ShellError> {
@@ -101,8 +102,8 @@ macro_rules! lazy_command {
         #[cfg(test)]
         mod $test {
             use super::*;
-            use crate::test::test_polars_plugin_command;
             use nu_protocol::ShellError;
+            use $crate::test::test_polars_plugin_command;
 
             #[test]
             fn test_examples() -> Result<(), ShellError> {
@@ -170,8 +171,8 @@ macro_rules! lazy_command {
         #[cfg(test)]
         mod $test {
             use super::*;
-            use crate::test::test_polars_plugin_command;
             use nu_protocol::ShellError;
+            use $crate::test::test_polars_plugin_command;
 
             #[test]
             fn test_examples() -> Result<(), ShellError> {
@@ -180,49 +181,3 @@ macro_rules! lazy_command {
         }
     };
 }
-
-// LazyReverse command
-// Expands to a command definition for reverse
-lazy_command!(
-    LazyReverse,
-    "polars reverse",
-    "Reverses the LazyFrame",
-    vec![Example {
-        description: "Reverses the dataframe.",
-        example: "[[a b]; [6 2] [4 2] [2 2]] | polars into-df | polars reverse",
-        result: Some(
-            NuDataFrame::try_from_columns(
-                vec![
-                    Column::new(
-                        "a".to_string(),
-                        vec![Value::test_int(2), Value::test_int(4), Value::test_int(6),],
-                    ),
-                    Column::new(
-                        "b".to_string(),
-                        vec![Value::test_int(2), Value::test_int(2), Value::test_int(2),],
-                    ),
-                ],
-                None
-            )
-            .expect("simple df for test should not fail")
-            .into_value(Span::test_data()),
-        ),
-    },],
-    reverse,
-    test_reverse
-);
-
-// LazyCache command
-// Expands to a command definition for cache
-lazy_command!(
-    LazyCache,
-    "polars cache",
-    "Caches operations in a new LazyFrame.",
-    vec![Example {
-        description: "Caches the result into a new LazyFrame",
-        example: "[[a b]; [6 2] [4 2] [2 2]] | polars into-df | polars reverse | polars cache",
-        result: None,
-    }],
-    cache,
-    test_cache
-);
